better error handling in the MDB2 container
git-svn-id: https://svn.php.net/repository/pear/packages/Mail_Queue/trunk@258970 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Queue/Container/mdb2.php
@@ -3,7 +3,7 @@
 // +----------------------------------------------------------------------+
 // | PEAR :: Mail :: Queue :: MDB2 Container                              |
 // +----------------------------------------------------------------------+
-// | Copyright (c) 1997-2004 The PHP Group                                |
+// | Copyright (c) 1997-2008 Lorenzo Alberton                             |
 // +----------------------------------------------------------------------+
 // | This source file is subject to version 3.0 of the PHP license,       |
 // | that is bundled with this package in the file LICENSE, and is        |
@@ -42,7 +42,7 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      * Reference to the current database connection.
      * @var object PEAR::MDB2 instance
      */
-    var $db;
+    var $db = null;
 
     /**
      * Table for sql database
@@ -59,9 +59,9 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
     // {{{ Mail_Queue_Container_mdb2()
 
     /**
-     * Contructor
+     * Constructor
      *
-     * Mail_Queue_Container_mMDB2:: Mail_Queue_Container_mdb2()
+     * Mail_Queue_Container_mdb2()
      *
      * @param mixed $options    An associative array of connection option.
      *
@@ -82,15 +82,64 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
         if (!empty($options['pearErrorMode'])) {
             $this->pearErrorMode = $options['pearErrorMode'];
         }
-        $this->db =& MDB2::connect($options['dsn'], true);
-        if (PEAR::isError($this->db)) {
+        $res = $this->_connect($options['dsn']);
+        if (PEAR::isError($res)) {
+            return $res;
+        }
+        $this->setOption();
+    }
+
+    // }}}
+    // {{{ _connect()
+
+    /**
+     * Connect to database by using the given DSN string
+     *
+     * @param mixed &$db DSN string | array | MDB2 object
+     *
+     * @return boolean|PEAR_Error on error
+     * @access private
+     */
+    function _connect(&$db)
+    {
+        if (is_object($db) && is_a($db, 'MDB2_Driver_Common')) {
+            $this->db = &$db;
+        } elseif (is_string($db) || is_array($db)) {
+            include_once 'MDB2.php';
+            $this->db =& MDB2::connect($db);
+        } elseif (is_object($db) && MDB2::isError($db)) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
                 'MDB2::connect failed: '. MDB2::errorMessage($this->db));
         } else {
-            $this->db->setFetchMode(MDB2_FETCHMODE_ASSOC);
+            return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
+                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
+                'The given dsn was not valid in file '. __FILE__ . ' at line ' . __LINE__);
         }
-        $this->setOption();
+        if (PEAR::isError($this->db)) {
+            return $this->db;
+        }
+        return true;
+    }
+
+    // }}}
+    // {{{ _checkConnection()
+
+    /**
+     * Check if there's a valid db connection
+     *
+     * @return boolean|PEAR_Error on error
+     */
+    function _checkConnection() {
+        if (!is_object($this->db) || !is_a($this->db, 'MDB2_Driver_Common')) {
+            $msg = 'MDB2::connect failed';
+            if (PEAR::isError($this->db)) {
+                $msg .= ': '.$this->db->getMessage();
+            }
+            return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
+                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__, $msg);
+        }
+        return true;
     }
 
     // }}}
@@ -104,13 +153,9 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function _preload()
     {
-        if (!is_object($this->db) || !is_a($this->db, 'MDB2_Driver_Common')) {
-            $msg = 'MDB2::connect failed';
-            if (PEAR::isError($this->db)) {
-                $msg .= ': '.$this->db->getMessage();
-            }
-            return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
-                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__, $msg);
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
         }
         $query = 'SELECT * FROM ' . $this->mail_table
                 .' WHERE sent_time IS NULL AND try_sent < '. $this->try
@@ -177,13 +222,9 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
     function put($time_to_send, $id_user, $ip, $sender,
                 $recipient, $headers, $body, $delete_after_send=true)
     {
-        if (!is_object($this->db) || !is_a($this->db, 'MDB2_Driver_Common')) {
-            $msg = 'MDB2::connect failed';
-            if (PEAR::isError($this->db)) {
-                $msg .= ': '.$this->db->getMessage();
-            }
-            return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
-                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__, $msg);
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
         }
         $id = $this->db->nextID($this->sequence);
         if (empty($id) || PEAR::isError($id)) {
@@ -209,7 +250,7 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
         if (PEAR::isError($res)) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
-                'MDB2::query failed - "'.$query.'" - '.MDB2::errorMessage($res));
+                'MDB2::query failed - "'.$query.'" - '.$res->getMessage());
         }
         return $id;
     }
@@ -226,6 +267,10 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function countSend($mail)
     {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
         if (!is_object($mail) || !is_a($mail, 'mail_queue_body')) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_UNEXPECTED, __FILE__, __LINE__);
         }
@@ -254,6 +299,10 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function setAsSent($mail)
     {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
         if (!is_object($mail) || !is_a($mail, 'mail_queue_body')) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_UNEXPECTED,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
@@ -284,6 +333,10 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function getMailById($id)
     {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
         $query = 'SELECT * FROM ' . $this->mail_table
                 .' WHERE id = '   . (int)$id;
         $row = $this->db->queryRow($query, null, MDB2_FETCHMODE_ASSOC);
@@ -321,6 +374,10 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function deleteMail($id)
     {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
         $query = 'DELETE FROM ' . $this->mail_table
                 .' WHERE id = ' . $this->db->quote($id, 'text');
         $res = $this->db->query($query);
